Made multiple changes to collisions, object loading, and positions.
Changed collision functionality, now more reliable & robust. Changed
X and Y in PositionComponent to represent the center of any entity
that has a collision component.

Objects now load from the map with a width and a height, and all
entities have a width and height different than that in RenderComponent.
This makes collisions tighter around the edges.