Stairs! Stairs! More stairs! Also implemented sprint (Spacebar).
Stairs now move you up and down floors, nice and smoothly.
There are probably some severe memory leaks there, but it doesn't
crash so that's good. Changed the map file loading, so the target
position stairs take you to is a float now, instead of an int, so
you can be at X.5 tiles, etc.

Hold space to sprint, this can be used for something else if we
want, it is implemented in movement_system, so it isn't hard to
change.